Move `getQueryString` helper function
We will use this to construct a professions sorting link that preserves filters, so move we move to a more generic location
